% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{hit_miss}
\alias{hit_miss}
\title{Monte-Carlo integration using the hit and miss method}
\usage{
hit_miss(ftn, a, b, f.min, f.max, n)
}
\arguments{
\item{ftn}{a function of single variable.}

\item{a}{lower bound of the integration. Must be finite.}

\item{b}{upper bound of the integration. Must be finite and assume b > a.}

\item{f.min}{lower bound of the ftn over the range [a,b].}

\item{f.max}{upper bound of the ftn over the range [a,b].}

\item{n}{number of samples used in the estimation.}
}
\description{
Monte-Carlo integration using the hit and miss method
}
\examples{
## We see that the number of repetitions n needs to be very large
## in order to get even just two decimal places of accuracy.
f <- function(x) return(x^3 - 7 * x^2 + 1 )
hit_miss(f, 0, 1, -6, 2, 1000)
hit_miss(f, 0, 1, -6, 2, 10000)
hit_miss(f, 0, 1, -6, 2, 100000)
hit_miss(f, 0, 1, -6, 2, 1000000)
}
\references{
Owen Jones,Robert Maillardet,Andrew Robinson (2014).Introduction to Scientific Programming and Simulation Using R ; Second Edition
}
\seealso{
\code{integrate}
}
