% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{hmplot}
\alias{hmplot}
\title{Plot Monte-Carlo integration using the hit and miss method}
\usage{
hmplot(ftn, a, b, c, d, n)
}
\arguments{
\item{ftn}{a function of single variable.}

\item{a}{lower bound of the integration. Must be finite.}

\item{b}{upper bound of the integration. Must be finite and assume b > a.}

\item{c}{lower bound of the ftn over the range [a,b].}

\item{d}{upper bound of the ftn over the range [a,b].}

\item{n}{number of samples used in the estimation.}
}
\description{
Partially vectorised version.
}
\examples{
## We have added a line to plot the successive approximations to the integral.
f <- function(x) return(x^3 - 7 * x^2 + 1 )
hmplot(f, 0, 1, -6, 2, 10000)
lines(c(0, 10000), c(-13/12, -13/12))
}
\references{
Owen Jones,Robert Maillardet,Andrew Robinson (2014).Introduction to Scientific Programming and Simulation Using R ; Second Edition
}
\seealso{
\code{plot}
}
