% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{trapezoid}
\alias{trapezoid}
\title{Trapezoidal Integration}
\usage{
trapezoid(ftn, a, b, n = 100)
}
\arguments{
\item{ftn}{a function of single variable.}

\item{a}{lower bound of the integration. Must be finite.}

\item{b}{upper bound of the integration. Must be finite and assume b > a.}

\item{n}{number of partitions.}
}
\description{
The numerical integration of ftn from a to b,using the trapezoidal rule with n subdivisions.
}
\details{
The integration calculates the trapezoid area of each subdivisions and sums them up.
}
\examples{
ftn1 <-  function(x) return(4 * x^3)
trapezoid(ftn1, 0, 1, n = 100)
}
\references{
Owen Jones,Robert Maillardet,Andrew Robinson (2014).Introduction to Scientific Programming and Simulation Using R ; Second Edition
}
\seealso{
\code{integrate}
}
