% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{trapezoid_adpt}
\alias{trapezoid_adpt}
\title{Adapted trapezoidal integration}
\usage{
trapezoid_adpt(ftn, a, b, tol = 1e-08, trace = FALSE)
}
\arguments{
\item{ftn}{a function of single variable.}

\item{a}{lower bound of the integration. Must be finite.}

\item{b}{upper bound of the integration. Must be finite and assume b > a.}

\item{tol}{the partition used is recursively refined until the estimate on successive partitions differs by at most tol.}

\item{trace}{if trace is TRUE then intermediate results are printed.}
}
\description{
The function returns a vector of length 2 whose first element is the integration and whose second element is the number of function evaluations required.
}
\details{
The variable tol is the tolerant error of the approximated integration and the partitions are subdivided.
t.recursion is the function that recursively improves the partitions such that the summation of the errors is less than the tolerant error.
}
\examples{
ftn1 <- function(x) return(4 * x^3)
trapezoid_adpt(ftn1, 0, 1, 0.00001, trace = FALSE)
}
\references{
Owen Jones,Robert Maillardet,Andrew Robinson (2014).Introduction to Scientific Programming and Simulation Using R ; Second Edition
}
\seealso{
\code{integrate}
}
