% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{trapezoid_rp}
\alias{trapezoid_rp}
\title{Trapezoidal Integration with random partitions}
\usage{
trapezoid_rp(ftn, a, b, n)
}
\arguments{
\item{ftn}{a function of single variable.}

\item{a}{lower bound of the integration. Must be finite.}

\item{b}{upper bound of the integration. Must be finite and assume b > a.}

\item{n}{number of partitions.}
}
\description{
Trapezoidal Integration with random partitions
}
\details{
The lengths of the partitions are not equal and the partitions are randomly selected by uniform distribution[a, b].
The uniform distribution has density
f(x) = 1/(max - min)
for min ≤ x ≤ max.
}
\examples{
## You can set.seed() before using this function in order
## to lead to the same result.
set.seed(1)
ftn1 <- function(x) return(4 * x^3)
trapezoid_rp(ftn1, 0, 1, n = 100)
}
\references{
Owen Jones,Robert Maillardet,Andrew Robinson (2014).Introduction to Scientific Programming and Simulation Using R ; Second Edition
}
\seealso{
\code{integrate}
}
